Move styleds.ts
