+++ src/components/swap/AdvancedSwapDetails.tsx
@@ -7,7 +7,7 @@ import { Field } from '../../state/swap/actions'
 import { CursorPointer, TYPE } from '../../theme'
 import { computeSlippageAdjustedAmounts, computeTradePriceBreakdown } from '../../util/prices'
 import { AutoColumn } from '../Column'
-import { SectionBreak } from '../ExchangePage/styleds'
+import { SectionBreak } from './styleds'
 import QuestionHelper from '../Question'
 import { RowBetween, RowFixed } from '../Row'
 import SlippageTabs, { SlippageTabsProps } from '../SlippageTabs'
