+++ src/components/swap/FormattedPriceImpact.tsx
@@ -1,7 +1,7 @@
 import { Percent } from '@uniswap/sdk'
 import React from 'react'
 import { ONE_BIPS } from '../../constants'
-import { ErrorText } from '../ExchangePage/styleds'
+import { ErrorText } from './styleds'
 import { warningServerity } from './PriceBar'
 
 export default function FormattedPriceImpact({ priceImpact }: { priceImpact?: Percent }) {
