+++ src/pages/Swap/index.tsx
@@ -19,7 +19,7 @@ import {
   StyledBalanceMaxMini,
   TruncatedText,
   Wrapper
-} from '../../components/ExchangePage/styleds'
+} from '../../components/swap/styleds'
 import QuestionHelper from '../../components/Question'
 import { AutoRow, RowBetween, RowFixed } from '../../components/Row'
 import { AdvancedSwapDetails } from '../../components/swap/AdvancedSwapDetails'
